requirements.txt
beautiful-date==2.3.0
gcsa==2.4.0
numpy==1.21.6
opencv-python==4.9.0.80
pytesseract==0.3.10
python-dotenv==0.21.1
